Add option: -keep-path.
When a file is explicity passed to the command-line:

  ocp-ocamlbuild src/mod.cmi

The filename was the complete path, including the slashes. By default,
the new behaviour is to keep only the basename (e.g. `mod.cmi`).

A new option `-keep-path` allow to keep the path as a hierarchical
structure (a.k.a. `OcamlRes.Res.Dir`) that can be merged with the
commandline-provided directories, e.g.;

  ocp-ocamlbuild -keep-path src/mod.cmi src/dir/